requirements.txt
selenium
sqlalchemy
psycopg2-binary
schedule
requests
python-dotenv
